Revert "APIS-7086 - Calling TPDS for sessions (#849)"
This reverts commit 2fc143dab7973143589fdfecad25ac2cc7db2edd.
