+++ conf/application.conf
@@ -109,7 +109,6 @@ feedbackBanner {
 
 fraudPreventionLink {
   enabled = false
-  
   apisWithFraudPrevention = [
     "vat-api",
     "business-details-api",
@@ -165,11 +164,6 @@ microservice {
     third-party-developer {
       host = localhost
       port = 9615
-    }
-
-    third-party-developer-session {
-      host = localhost
-      port = 9616
     }
 
     third-party-application-production {
