+++ run_local_with_dependencies.sh
@@ -4,6 +4,6 @@ sm2 -start MONGO DATASTREAM
 
 sm2 -start EMAIL HMRC_EMAIL_RENDERER DIGITAL_CONTACT_STUB DESKPRO_STUB DESKPRO_TICKET_QUEUE
 
-sm2 -start THIRD_PARTY_APPLICATION THIRD_PARTY_DEVELOPER THIRD_PARTY_DEVELOPER_SESSION API_SUBSCRIPTION_FIELDS API_DEFINITION THIRD_PARTY_DELEGATED_AUTHORITY API_PLATFORM_EVENTS API_PLATFORM_MICROSERVICE THIRD_PARTY_ORCHESTRATOR PUSH_PULL_NOTIFICATIONS_API API_PLATFORM_XML_SERVICES PUSH_PULL_NOTIFICATIONS_GATEWAY
+sm2 -start THIRD_PARTY_APPLICATION THIRD_PARTY_DEVELOPER API_SUBSCRIPTION_FIELDS API_DEFINITION THIRD_PARTY_DELEGATED_AUTHORITY API_PLATFORM_EVENTS API_PLATFORM_MICROSERVICE THIRD_PARTY_ORCHESTRATOR PUSH_PULL_NOTIFICATIONS_API API_PLATFORM_XML_SERVICES PUSH_PULL_NOTIFICATIONS_GATEWAY
 
 ./run_local.sh
