+++ test/uk/gov/hmrc/thirdpartydeveloperfrontend/controllers/endpointauth/EndpointScenarioSpec.scala
@@ -86,7 +86,6 @@ abstract class EndpointScenarioSpec extends AsyncHmrcSpec with GuiceOneAppPerSui
     GuiceApplicationBuilder()
       .overrides(bind[ConnectorMetrics].to[NoopConnectorMetrics])
       .overrides(bind[ThirdPartyDeveloperConnector].toInstance(tpdConnector))
-      .overrides(bind[ThirdPartyDeveloperSessionConnector].toInstance(tpdSessionConnector))
       .overrides(bind[ThirdPartyApplicationProductionConnector].toInstance(tpaProductionConnector))
       .overrides(bind[ThirdPartyApplicationSandboxConnector].toInstance(tpaSandboxConnector))
       .overrides(bind[DeskproConnector].toInstance(deskproConnector))
@@ -207,8 +206,8 @@ abstract class EndpointScenarioSpec extends AsyncHmrcSpec with GuiceOneAppPerSui
   when(tpdConnector.getOrCreateUserId(*[LaxEmailAddress])(*)).thenReturn(Future.successful(UserId.random))
   when(tpdConnector.reset(*)(*)).thenReturn(Future.successful(OK))
   when(tpdConnector.authenticate(*)(*)).thenReturn(Future.successful(UserAuthenticationResponse(false, false, None, Some(session))))
-  when(tpdSessionConnector.fetchSession(eqTo(sessionId))(*)).thenReturn(Future.successful(session))
-  when(tpdSessionConnector.deleteSession(eqTo(sessionId))(*)).thenReturn(Future.successful(OK))
+  when(tpdConnector.fetchSession(eqTo(sessionId))(*)).thenReturn(Future.successful(session))
+  when(tpdConnector.deleteSession(eqTo(sessionId))(*)).thenReturn(Future.successful(OK))
   when(tpdConnector.authenticateMfaAccessCode(*)(*)).thenReturn(Future.successful(session))
   when(tpdConnector.verify(*)(*)).thenReturn(Future.successful(OK))
   when(tpaProductionConnector.verify(*)(*)).thenReturn(Future.successful(ApplicationVerificationSuccessful))
